have the user confirm they definitely want to disconnect an identity, before doing so (#361)

+++ app/routes/settings/connected-accounts.js
@@ -33,6 +33,13 @@ export default Route.extend(AuthenticatedRouteMixin, EnsureStatefulLogin, {
         identityProvider = 'Email/Password';
       }
 
+      let confirmation = confirm(
+        `Are you sure you want to disconnect your ${identityProvider} account?`,
+      );
+      if (!confirmation) {
+        return;
+      }
+
       identity
         .destroyRecord()
         .then(() => {
